requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
imageio==2.20.0
numpy==1.22.3
opencv_python==4.6.0.66
Pillow==9.3.0
scikit_learn==1.1.3
scipy==1.8.1
six==1.16.0
tensorflow-cpu==2.8.0

click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
